requirements.txt
streamlit==1.31
scikit-learn==1.4.0
thefuzz==0.22.1
numpy==1.26
pandas==2.1
matplotlib==3.8
python-dotenv==1.0.1
voyageai==0.1.7
